modifications induced by change in "head_version" option of gaze controller
svn path=/trunk/iCub/; revision=26243

+++ src/modules/actionsRenderingEngine/src/MotorThread.cpp
@@ -2472,11 +2472,11 @@ bool MotorThread::exploreTorso(Bottle &options)
 
     Bottle info;
     ctrl_gaze->getInfo(info);
-    int head_version=info.check("head_version",Value(1)).asInt();
+    double head_version=info.check("head_version",Value(1.0)).asDouble();
 
     ostringstream type;
     type<<(dominant_eye==LEFT?"left":"right");
-    if (head_version==2)
+    if (head_version==2.0)
         type<<"_v2";
 
     iCubEye iKinTorso=iCubEye(type.str());
